Implemented district field support and enhanced offline/search:
- Added district column to businesses table
- Introduced district filtering in search_businesses and Discover UX
- Enabled district input on registration/registration flows
- Updated related sales/stock pages to handle decimal quantities and custom pricing
- Added UI hints and validations for district data
- Included migration to create district field and search function respecting district filter

X-Lovable-Edit-ID: edt-1069ff05-c0a1-4fe6-a708-5fb6d5b2e103

+++ src/pages/PurchasesPage.tsx
@@ -65,7 +65,7 @@ export default function PurchasesPage() {
       category: toSentenceCase(form.category.trim()),
       quality: toSentenceCase(form.quality.trim()),
       unit_type: form.unit_type,
-      quantity: parseInt(form.quantity) || 1,
+      quantity: parseFloat(form.quantity) || 1,
       unit_price: parseFloat(form.unit_price) || 0,
       wholesale_price: parseFloat(form.wholesale_price) || parseFloat(form.unit_price) || 0,
       retail_price: parseFloat(form.retail_price) || parseFloat(form.unit_price) || 0,
